# Fix slots usage and use more slots