Use empty slots for ABC classes, otherwise classes which inherit from them will still have `__dict__`. Also use `__slots__` for more classes.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/dpkp/kafka-python/1987)
<!-- Reviewable:end -->