@carsonip 2020-01-17
This saves ~200 bytes per object, because the object does not contain a `__dict__`, and there are also a marginal performance gain:
```
python benchmarks/record_batch_read.py
```
before:
```
.....................
batch_read_v0: Mean +- std dev: 1.23 ms +- 0.06 ms
.....................
batch_read_v1: Mean +- std dev: 1.28 ms +- 0.10 ms
.....................
batch_read_v2: Mean +- std dev: 961 us +- 82 us
```
after:
```
.....................
batch_read_v0: Mean +- std dev: 1.18 ms +- 0.05 ms
.....................
batch_read_v1: Mean +- std dev: 1.20 ms +- 0.06 ms
.....................
batch_read_v2: Mean +- std dev: 950 us +- 55 us
```